add flower.center.core
