+++ flower.center/flower.center.core/pom.xml
@@ -20,15 +20,19 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>com.ly.train</groupId>
-        <artifactId>flower</artifactId>
+        <artifactId>flower.center</artifactId>
         <version>0.1.3</version>
-        <relativePath>..</relativePath>
     </parent>
-    <artifactId>flower.center</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>flower.center.api</module>
-        <module>flower.center.impl</module>
-        <module>flower.center.core</module>
-    </modules>
+    <artifactId>flower.center.core</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.ly.train</groupId>
+            <artifactId>flower.center.api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.ly.train</groupId>
+            <artifactId>flower.web</artifactId>
+        </dependency>
+    </dependencies>
 </project>
