+++ flower.center/flower.center.core/src/main/java/com/ly/train/flower/center/core/controller/serviceconfig/ServiceConfigListController.java
@@ -16,39 +16,39 @@
 /**
  * 
  */
-package com.ly.train.flower.center.controller.serviceinfo;
+package com.ly.train.flower.center.core.controller.serviceconfig;
 
 import java.io.IOException;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.alibaba.fastjson.JSONObject;
-import com.ly.train.flower.center.controller.BaseController;
-import com.ly.train.flower.center.service.ReturnService;
-import com.ly.train.flower.center.service.ServiceInfoListService;
+import com.ly.train.flower.center.core.controller.BaseController;
+import com.ly.train.flower.center.core.service.ReturnService;
+import com.ly.train.flower.center.core.service.ServiceConfigListService;
 import com.ly.train.flower.common.annotation.Flower;
-import com.ly.train.flower.registry.config.ServiceInfo;
+import com.ly.train.flower.common.service.config.ServiceConfig;
 
 /**
  * @author leeyazhou
  * 
  */
 @RestController
-@RequestMapping("/service/")
-@Flower(value = "listServiceInfo")
-public class ServiceInfoListController extends BaseController {
+@RequestMapping("/serviceconfig/")
+@Flower(serviceName = "ServiceA", value = "ServiceA", flowNumber = 32)
+public class ServiceConfigListController extends BaseController {
 
   @RequestMapping("list")
   protected void process(String data, HttpServletRequest req) throws IOException {
-    ServiceInfo serviceInfo = JSONObject.parseObject(data, ServiceInfo.class);
-    logger.info("请求参数 : {}", serviceInfo);
-    doProcess(serviceInfo, req);
+    ServiceConfig serviceConfig = JSONObject.parseObject(data, ServiceConfig.class);
+    logger.info("请求参数 : {}", serviceConfig);
+    doProcess(serviceConfig, req);
   }
 
 
 
   @Override
   public void buildFlower() {
-    getServiceFlow().buildFlow(ServiceInfoListService.class, ReturnService.class);
+    getServiceFlow().buildFlow(ServiceConfigListService.class, ReturnService.class);
   }
 }
