+++ flower.center/flower.center.core/src/main/java/com/ly/train/flower/center/core/service/ServiceInfoListService.java
@@ -16,7 +16,7 @@
 /**
  * 
  */
-package com.ly.train.flower.center.service;
+package com.ly.train.flower.center.core.service;
 
 import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
