+++ flower.center/flower.center.impl/src/main/resources/application.yml
@@ -23,7 +23,4 @@ spring:
          
 flower:
    center:
-      redis:
-         host: 10.100.216.147
-         port: 6399
-         password: flowerpassword
+      store: memory
